Datasets/evalplus/Java/src/test/java/com/example/HumanEval_69Test.java
package com.example;

import org.junit.Test;
import static org.junit.Assert.assertEquals;

public class HumanEval_69Test {
    
    @Test
    public void test0() {
        assertEquals(1, HumanEval_69.search(new int[]{5, 5, 5, 5, 1}));
    }

    @Test
    public void test1() {
        assertEquals(4, HumanEval_69.search(new int[]{4, 1, 4, 1, 4, 4}));
    }

    @Test
    public void test2() {
        assertEquals(-1, HumanEval_69.search(new int[]{3, 3}));
    }

    @Test
    public void test3() {
        assertEquals(8, HumanEval_69.search(new int[]{8, 8, 8, 8, 8, 8, 8, 8}));
    }

    @Test
    public void test4() {
        assertEquals(2, HumanEval_69.search(new int[]{2, 3, 3, 2, 2}));
    }

    // Continue writing the rest of the tests in the same way...

    @Test
    public void test33() {
        assertEquals(3, HumanEval_69.search(new int[]{1, 1, 11, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3}));
    }
}

// End of Test
